fix: fix sheets numfmt panel
