+++ packages/sheets-numfmt/src/index.ts
@@ -32,7 +32,5 @@ export { SetCurrencyCommand } from './commands/commands/set-currency.command';
 export { type ISetNumfmtCommandParams, SetNumfmtCommand } from './commands/commands/set-numfmt.command';
 export { SetPercentCommand } from './commands/commands/set-percent.command';
 export { SubtractDecimalCommand } from './commands/commands/subtract-decimal.command';
-export { CloseNumfmtPanelOperator } from './commands/operations/close.numfmt.panel.operation';
-export { OpenNumfmtPanelOperator } from './commands/operations/open.numfmt.panel.operation';
 
 // #endregion
