+++ packages/sheets-numfmt/tsconfig.json
@@ -5,5 +5,5 @@
         "outDir": "lib/types"
     },
     "references": [{ "path": "./tsconfig.node.json" }],
-    "include": ["src"]
+    "include": ["src", "../sheets-numfmt-ui/src/commands/operations"]
 }
